requirements.txt
kopf[dev]
kubernetes
pydantic
mysql-connector-python==8.0.26 # issue with 8.0.27 and Azure
hvac
pyyaml
twine
wheel
pylint
pwgen